[重构日志系统并优化CI/CD配置]: 提升构建流程与日志处理效率
- 更新 `.clang-tidy` 配置，细化代码检查规则并启用更多检查项，提升代码质量分析能力
- 优化 GitHub Actions 工作流，移除冗余的 `os_name` 输入参数，简化缓存键配置
- 修正 Dependabot 配置，正确指定目录路径以支持依赖更新
- 新增 CodeQL 分析工作流，定期扫描代码安全漏洞并忽略无关文件变更
- 升级 CMake 配置，将 C++ 标准提升至 C++20 并添加编译器版本检查
- 调整 Qt 配置，更新宏定义以禁用过时的 Qt 功能
- 重构日志模块，将文件操作逻辑拆分至独立的 `LogFile` 类，优化日志输出性能
- 修改日志模块枚举类定义，使用位掩码支持组合选项，增强可读性和扩展性
- 调整日志输出逻辑，控制控制台日志长度并优化文件写入效率
- 修正多个源文件中的捕获列表，使用更合适的捕获方式替代 lambda 捕获
- 修正 FFmpeg 相关代码中的数据类型转换问题，避免潜在的类型溢出风险
- 统一项目描述信息，将项目名称更新为 "Qt-Media" 以反映项目实际用途
- 修正日志初始化代码，使用正确的枚举值配置日志输出方向
- 更新 CI/CD 工作流的缓存策略，提高构建效率并减少冗余操作

+++ src/ffmpeg/frame.cc
@@ -235,7 +235,7 @@ auto Frame::fromQImage(const QImage &image) -> Frame *
     avFrame->format = AV_PIX_FMT_RGBA;
     framePtr->getBuffer();
 
-    memcpy(avFrame->data[0], img.bits(), avFrame->width * avFrame->height * 4);
+    memcpy(avFrame->data[0], img.bits(), static_cast<size_t>(avFrame->width) * avFrame->height * 4);
 
     return framePtr.release();
 }
